requirement.txt
streamlit
torch
transformers
Pillow


